requirements.txt
blinker==1.9.0
click==8.2.2
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.2
SQLAlchemy==2.0.42
typing_extensions==4.14.1
Werkzeug==3.1.3
